+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Characteristic Changelog
 
+## 1.0.0-beta.6 - Unreleased
+### Fixed
+- Errors caused by soft-deleting and restoring characteristic elements
+- Fixed Drilldown not ignoring deleted characteristics
+
 ## 1.0.0-beta.5 - 3/24/20
 ### Fixed
 - Issue where non-admin users could not edit characteristics